split cmakelists, updated readme
