+++ examples/CMakeLists.txt
@@ -1,11 +1,11 @@
 if(${CMAKE_SOURCE_DIR} STREQUAL ${CMAKE_CURRENT_SOURCE_DIR})
     # Settings when 'examples' is the root projet
-    cmake_minimum_required(VERSION 3.14)
+    cmake_minimum_required(VERSION 3.20)
     project(zenohcpp_examples LANGUAGES C CXX)
     include(../cmake/helpers.cmake)
     set_default_build_type(Release)
-    configure_include_project(ZENOHCPP_ZENOHC zenohc zenohc::lib "../../zenoh-c" zenohc "https://github.com/eclipse-zenoh/zenoh-c")
-    configure_include_project(ZENOHCPP zenohcpp zenohcpp ".." zenohcpp "https://github.com/eclipse-zenoh/zenoh-cpp")
+    configure_include_project(ZENOHCPP_ZENOHC zenohc zenohc::lib "../../zenoh-c" zenohc "https://github.com/eclipse-zenoh/zenoh-c" "")
+    configure_include_project(ZENOHCPP zenohcpp zenohcpp ".." zenohcpp "https://github.com/eclipse-zenoh/zenoh-cpp" "")
     add_custom_target(examples ALL)
 else()
     message(STATUS "zenoh-cpp examples")
@@ -26,7 +26,6 @@ function(add_example file suffix is_cpp)
 		set_property(TARGET ${target} PROPERTY LANGUAGE C)
 		set_property(TARGET ${target} PROPERTY C_STANDARD 11)
 	endif()
-	set_property(TARGET ${target} PROPERTY RUNTIME_OUTPUT_DIRECTORY "${CMAKE_CURRENT_BINARY_DIR}")
 	add_platform_libraries(${target})
 endfunction()
