+++ include/zenohcpp.h
@@ -13,7 +13,7 @@
 
 #pragma once
 
-#include "zenohcpp_base.h"
-#include "zenohcpp_channel.h"
-#include "zenohcpp_objects.h"
-#include "zenohcpp_structs.h"
+#include "zenohcpp/zenohcpp_base.h"
+#include "zenohcpp/zenohcpp_channel.h"
+#include "zenohcpp/zenohcpp_objects.h"
+#include "zenohcpp/zenohcpp_structs.h"
